Добавляет сортировку миниатюр
